[Chore] #371 - SendThumnailURLDelegate 및 SendSparkCellDelegate 프로토콜 분리
SendThumnailURLDelegate 및 SendSparkCellDelegate 프로토콜 분리
