+++ Spark-iOS/Spark-iOS/Source/Protocols/SendSparkCellDelegate.swift
@@ -1,0 +1,13 @@
+//
+//  SendSparkCellDelegate.swift
+//  Spark-iOS
+//
+//  Created by Junho Lee on 2022/03/12.
+//
+
+import Foundation
+
+protocol SendSparkCellDelegate: AnyObject {
+    func sendSpark(_ content: String)
+    func showTextField()
+}
