+++ Spark-iOS/Spark-iOS/Source/Protocols/SendThumbnailURLDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  SendThumbnailURLDelegate.swift
+//  Spark-iOS
+//
+//  Created by Junho Lee on 2022/03/12.
+//
+
+import Foundation
+
+protocol SendThumbnailURLDelegate: AnyObject {
+    func sendThumbnailURL(url: String)
+}
